% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ivt.R
\name{theme_ivt}
\alias{theme_ivt}
\title{Sligth tweaks to \code{theme_light}}
\usage{
theme_ivt()
}
\description{
Sligth tweaks to \code{theme_light}
}
\examples{
ggplot(iris, aes(x = Sepal.Width, y = Sepal.Length)) +
  geom_point() +
  theme_ivt()
}
